fix ZLogs
